Create String slice operator
The slice operator can be used to print from a given index to another index. The below code prints from 3rd to 6th character of the string value collegename. 
(As the index starts from zero, to print the 3rd character the index will be 2). 
collegename = input() print(collegename[2:6])
Complete the program so that it prints the initial 5 characters of the input string value. (Assume that the length of the input string is more than 5).