Create String Slice operator continued
The slice operator can be used to print from a given index till the end of the string. 
The below program prints from 3rd character till the end of the string value name. 
Please NOTE that there is no index after : name = input() print(name[2:])
Complete the program so that it prints the input string valueA from the 2nd character to the end of the input string value.