Create String upper or lower case
upper and lower methods can be used to convert all the characters in a string to upper case or lower case respectively. 
The below program accepts a string value and prints all characters in upper case. inputstring = input() print (inputstring.upper())
Complete the program so that accepts a string value and prints all characters in lower case.