Create String - reverse
Reversing the string can be done by using the slice operator to go one step back from the end by specifying -1. 
The below program accepts a string value and 
prints the reverse value. inputstring = input() print (inputstring[::-1])
Complete the program so that accepts two string values, concatenates them and then prints the reverse of the concatenated value as the output.