Create String replace
replace method helps in replacing a specific string value (one or more characters) with another string value. 
As an example, the below code replaces a with z in the input string value. (If the input is test, the ouput is zesz). 
inputstring = input() print (inputstring.replace("t","z"))
Complete the program so that accepts an input string value and replaces j with y and prints the replaced value as the output.