Create String swapcase
swapcase method helps in changing upper case letters to lower case and lower case letters to upper case. 
As an example, if the input to the program given below is AbCDeF, then the output is aBcdEf inputstr = input() print (inputstr.swapcase())
Complete the program so that accepts an input string value and swaps the case from the 3rd character. (Assume the input string is of length > 3)