Create String - Join
join method helps in returning a string where the string values in an input sequence are joined by a separator. As an example, if the input to the program given below is abc, the output is a-b-c inputstring = input() sep="-" print (sep.join(inputstring))
Complete the program so that accepts an input string value and separates the characters in the string using colon. - If the input is aeiou, the output must be a:e:i:o:u