Create Space Separated String Input
A sentence S is passed as input with one or more spaces in between the words in it. The program must accept S and print it as the output. 
Boundary Conditions: 1 <= Length of S <= 100 
Example Input/Output 1: 
Input: Good Morning. Please have a cup of coffee. 
Output: Good Morning. Please have a cup of coffee.