Create MFIB - Sum of Integers - List
Fill in the blanks with code so that the program accepts a list of integers and prints their sum as the output. 
Boundary Condition(s): 1 <= Each integer value <= 1000 
Input Format: The first line contains the list of integers separated by a space. 
Output Format: The first line contains the sum of the integers in the list. 
Example Input/Output 1: Input: 5 4 2 10 1 Output: 22 
Example Input/Output 2: Input: 17 23 6 2 20 10 30 Output: 108