Create MFIB - Sum of Minimum & Maximum
Fill in the blanks with code so that the program accepts a list of integers and prints the sum of the minimum integer and the maximum integer in the list. 
Boundary Condition(s): 1 <= Each integer value <= 10^8 
Input Format: The first line contains the list of integers separated by a space. 
Output Format: The first line contains the sum of the minimum integer and the maximum integer in the list. 
Example Input/Output 1: Input: 7 5 9 2 Output: 11 
Example Input/Output 2: Input: 67 89 34 99 10 56 Output: 109