Create MFIB - Integers in the Odd Positions
Fill in the blanks with code so that the program accepts a list of integers and prints the integers in the odd positions as the output. 
Boundary Condition(s): 1 <= Each integer value <= 1000 
Input Format: The first line contains the list of integers separated by a space. 
Output Format: The first line contains the integers in the odd positions in the list separated by a space. 
Example Input/Output 1: Input: 16 25 7 12 8 11 Output: 16 7 8 Example Input/Output 2: Input: 5 23 6 2 9 17 30 Output: 5 6 9 30