Create MFIB - Search the Integer X
The program must accept a list of integers and an integer X as the input. The program must print YES if X is present in the given list of integers. 
Else the program must print NO as the output. 
Please fill in the blanks with code so that the program runs successfully. 
Input Format: The first line contains a list of integers separated by a space. 
Output Format: The first line contains either YES or NO. 
Example Input/Output 1: Input: 10 21 8 15 62 18 4 36 15 Output: YES
Example Input/Output 2: Input: 51 16 11 34 8 19 12 Output: NO