Create MFIB - Exclude First Two Integers
Fill in the blanks with code so that the program accepts a list of integers and prints the integers from the third integer as the output. Note: The number of integers in the list is always greater than 2. 
Boundary Condition(s): 1 <= Each integer value <= 1000 
Input Format: The first line contains the list of integers separated by a space. 
Output Format: The first line contains the integers from the third integer separated by a space. 
Example Input/Output 1: Input: 8 4 5 19 30 37 11 21 Output: 5 19 30 37 11 21 
Example Input/Output 2: Input: 17 2 12 5 67 8 Output: 12 5 67 8