Create Python - Arithmetic Operators
The arithmetic operators in Python are + - * / for addition, subtraction, multiplication and division respectively. % is for modulus (remainder). ** is for exponent (power of). // if for floor division. (Example: 9.0//2.0 is equal to 4.0). 
The below program prints 194.0 as the output. a = 200 b = 2 c = 20 d = 10 e = 4 print (a+b-c/d*e)
print(a + c / b - d / e)  # This will print 185.0
no this code output is 185.5