Create El sum of First and last
Rectify the error(s) so that the program accepts a list of integers and prints the sum of the first and the last integer in the list as the output. Note: The number of integers in the list is always greater than 1. 
Boundary Condition(s): 1 <= Each integer value <= 1000 
Input Format: The first line contains the list of integers separated by a space. 
Output Format: The first line contains the sum of the first and the last integer in the list. 
Example Input/Output 1: Input: 8 4 5 19 30 37 11 21 Output: 29 Example Input/Output 2: Input: 17 2 12 5 67 8 Output: 25