Create MFIB - Print from B to A - Reverse
Fill in the blanks with code so that the program accepts two integers (A, B) and prints the integers from B to A in reverse order. 
Boundary Condition(s): 1 <= A < B <= 100 
Input Format: The first line contains A and B separated by a space. 
Output Format: The first line contains the integers from B to A in reverse order separated by a space. 
Example Input/Output 1: Input: 3 8 Output: 8 7 6 5 4 3 
Example Input/Output 2: Input: 1 6 Output: 6 5 4 3 2 1