Create MFIB - All Unique or Not
Fill in the blanks with code so that the program accepts a list of integers and prints YES if all the integers in the list are unique. Else the program prints NO as the output. Boundary Condition(s): 1 <= Each integer value <= 10^8 
Input Format: The first line contains the list of integers separated by a space. 
Output Format: The first line contains YES or NO. 
Example Input/Output 1: Input: 7 8 9 3 Output: YES 
Example Input/Output 2: Input: 90 87 65 56 87 Output: NO