Create MFIB - Print till 10
The program must accept a list of integers as the input. The program must print the integers till the first occurrence of 10 in the given list as the output. Please fill in the blanks with code so that the program runs successfully. 
Note: At least one 10 is always present in the list of integers. Input Format: The first line contains a list of integers separated by a space. 
Output Format: The first line contains the integer(s) separated by a space. 
Example Input/Output 1: Input: 41 17 23 62 10 37 17 19 Output: 41 17 23 62 10 
Example Input/Output 2: Input: 64 10 98 10 48 10 Output: 64 10