Create MFIB - Time to Minutes
Fill in the blanks with code so that the program must accept the hour HH and the minutes MM separated by a colon as the input. The program must print the total minutes equivalent to the given hour and minutes as the output. 
Boundary Condition(s): 0 <= HH <= 23 0 <= MM <= 59 
Input Format: The first line contains the hour and minutes separated by a colon. 
Output Format: The first line contains the total minutes equivalent to the given hour and minutes.
Example Input/Output 1: Input: 2:30 Output: 150 
Example Input/Output 2: Input: 15:15 Output: 915