Create MFIB - Count of Unique Integers
Fill in the blanks with code so that the program accepts a list of integers and prints the count of unique integers in the list. 
Boundary Condition(s): 1 <= Each integer value <= 10^8 
Input Format: The first line contains the list of integers separated by a space. 
Output Format: The first line contains the sum of the count of unique integers in the list. 
Example Input/Output 1: Input: 8 7 2 1 7 4 2 1 Output: 5 
Example Input/Output 2: Input: 17 17 34 29 Output: 3