Create Modulo Operator
Fill in the missing lines of code to print the remainder when the input number is divided by 9.
As % is used for finding the remainder, the below program prints the remainder when the input number is divided by 5. 
If the input is 102, the below program prints 2 as the 
output. 
num = int(input()) 
print (num%5)