Create Floor Operator
Fill in the missing lines of code to print the first and middle letters of the odd length input string. 
Example Input/Output 1: Input: CRYBABY Output: CB 
Example Input/Output 2: Input: MANGO Output: MN.
As // is used as the floor operator, the below program prints the middle letter of a string of odd length. If the input is MINDSET, the below program prints D as the output. (As the index starts from zero and 7//2 is 3) 
inputstr = input() 
inputlength = len(inputstr) 
print (inputstr[inputlength//2])