Create if statement only Odd
Fill in the missing lines of code to print the input number only if it is odd.
The if statement contains a logical expression followed by a : The statements after the : must be uniformly indented. Else Python interpreter will throw an error. In Python, any non-zero and non-null value is considered True. Any zero or null value is considered False. 
The below program shows the use of if statement to print the number if it is an even number. 
The == operator checks for equality. 
num = int(input()) 
if num%2 == 0: 
print (num)