Create Python - Print SENIOR or NOT
Fill in the missing lines of code to print SENIOR if the input value age is more than 59. Else print NOTSENIOR. Note: The output is CASE SENSITIVE.
The else statement can be used along with an if statement. The statements in the else block are executed if the conditional expression in the if evaluates to False. The statements after the : in the else must be uniformly indented. 
Else Python interpreter will throw an error. 
The below program shows the use of if else statement to print odd or even depending on the input value. 
num = int(input()) 
if num%2 == 0: 
 print ("even") 
else: 
 print ("odd")