Create If not multiple of 5 add 10
Fill in the missing lines so that the program prints the input value N without modification when N is a multiple of 5. Else the program prints N+10 as the output. 
Example Input/Output 1: Input: 25 Output: 25 
Example Input/Output 2: Input: 222 Output: 232
== is used to check if two operands are equal. != is used to check if two operands are NOT equal. 
Hence the program to print odd or even based on the input number can be written either as 
num = int(input()) 
if num%2 == 0: 
print ("even") 
else: print ("odd") or as 
num = int(input()) 
if num%2 != 0: 
print ("odd") else: 
print ("even")