Create List to Set conversion
Fill in the missing lines of code so that the program below prints 1 2 3 4 5 as the output.
A set can be converted to a List using list() method. The below program will print 1 2 3 4 5 as output. 
numbersset = {1,2,4,5,3,5,4,3,2,1,2} 
numberslist = list(numbersset) 
numberslist.sort() 
for n in numberslist: 
print (n, end=" ")
In a similar fashion, a List can be converted to Set by using set() method.